Conflicts:
	recovery-coreextension-plugin/src/main/java/es/pfsgroup/recovery/ext/impl/asunto/model/EXTAsunto.java
	sql/9.1/producto/DML_557_ENTITY01_NORMALIZAR_VALORES_TERMINOS.sql
	sql/tool/templates/setEnvGlobalHAYA.sh

Former-commit-id: 7673d88236a94c5e5255c37bdab737d0ef07f528

+++ rec-common/src/main/java/es/capgemini/pfs/asunto/dao/impl/AsuntoDaoImpl.java
@@ -380,10 +380,10 @@ public class AsuntoDaoImpl extends AbstractEntityDao<Asunto, Long> implements As
         //MAX MINS
         if (requiereProcedimiento(dto)) {
             if (dto.getMaxSaldoTotalContratos() == null) {
-                dto.setMaxSaldoTotalContratos((double) Integer.MAX_VALUE);
+                dto.setMaxSaldoTotalContratos((float) Integer.MAX_VALUE);
             }
             if (dto.getMinSaldoTotalContratos() == null) {
-                dto.setMinSaldoTotalContratos(0d);
+                dto.setMinSaldoTotalContratos(0f);
             }
             if (dto.getMaxImporteEstimado() == null) {
                 dto.setMaxImporteEstimado((double) Integer.MAX_VALUE);
