+++ rec-common/src/main/java/es/capgemini/pfs/asunto/dto/DtoBusquedaAsunto.java
@@ -22,8 +22,8 @@ public class DtoBusquedaAsunto extends PaginationParamsImpl {
     private String comboEstados;
     private String comboSupervisor;
     private String estadoAnalisis;
-    private Double minSaldoTotalContratos;
-    private Double maxSaldoTotalContratos;
+    private Float minSaldoTotalContratos;
+    private Float maxSaldoTotalContratos;
     private Double minImporteEstimado;
     private Double maxImporteEstimado;
     private String filtroContrato;
@@ -173,28 +173,28 @@ public class DtoBusquedaAsunto extends PaginationParamsImpl {
     /**
      * @return the minSaldoTotalContratos
      */
-    public Double getMinSaldoTotalContratos() {
+    public Float getMinSaldoTotalContratos() {
         return minSaldoTotalContratos;
     }
 
     /**
      * @param minSaldoTotalContratos the minSaldoTotalContratos to set
      */
-    public void setMinSaldoTotalContratos(Double minSaldoTotalContratos) {
+    public void setMinSaldoTotalContratos(Float minSaldoTotalContratos) {
         this.minSaldoTotalContratos = minSaldoTotalContratos;
     }
 
     /**
      * @return the maxSaldoTotalContratos
      */
-    public Double getMaxSaldoTotalContratos() {
+    public Float getMaxSaldoTotalContratos() {
         return maxSaldoTotalContratos;
     }
 
     /**
      * @param maxSaldoTotalContratos the maxSaldoTotalContratos to set
      */
-    public void setMaxSaldoTotalContratos(Double maxSaldoTotalContratos) {
+    public void setMaxSaldoTotalContratos(Float maxSaldoTotalContratos) {
         this.maxSaldoTotalContratos = maxSaldoTotalContratos;
     }
